TML/requirementsconda.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=0.11.0=py38haa244fe_0
aiohttp=3.7.3=py38h294d835_1
async-timeout=3.0.1=py_1000
attrs=20.3.0=pyhd3deb0d_0
blas=1.0=mkl
blinker=1.4=py_1
brotlipy=0.7.0=py38h294d835_1001
ca-certificates=2021.1.19=haa95532_0
cachetools=4.1.1=py_0
certifi=2020.12.5=py38haa95532_0
cffi=1.14.4=py38hd8c33c5_1
chardet=3.0.4=py38h9bdc248_1008
click=7.1.2=pyh9f0ad1d_0
cryptography=3.3.1=py38hd8c33c5_1
cudatoolkit=10.2.89=h74a9793_1
cycler=0.10.0=py_2
freetype=2.10.4=hd328e21_0
google-auth=1.24.0=pyhd3deb0d_0
google-auth-oauthlib=0.4.1=py_2
grpcio=1.34.1=py38he5377a8_0
hdf5=1.10.4=h7ebc959_0
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha925a31_3
idna=2.10=pyh9f0ad1d_0
importlib-metadata=3.4.0=py38haa244fe_0
intel-openmp=2020.2=254
joblib=0.17.0=py_0
jpeg=9b=vc14h4d7706e_1
kiwisolver=1.3.1=py38hbd9d945_1
libopencv=4.0.1=hbb9e17c_0
libpng=1.6.37=h2a8f88b_0
libprotobuf=3.14.0=h7755175_0
libtiff=4.1.0=h56a325e_1
libuv=1.40.0=he774522_0
lz4-c=1.9.2=hf4a77e7_3
markdown=3.3.3=pyh9f0ad1d_0
matplotlib=3.3.3=py38haa244fe_0
matplotlib-base=3.3.3=py38h34ddff4_0
mkl=2020.2=256
mkl-service=2.3.0=py38h196d8e1_0
mkl_fft=1.2.0=py38h45dec08_0
mkl_random=1.1.1=py38h47e9c7a_0
multidict=5.1.0=py38h294d835_1
ninja=1.10.2=py38h6d14046_0
numpy=1.19.2=py38hadc3359_0
numpy-base=1.19.2=py38ha3acd2a_0
oauthlib=3.0.1=py_0
olefile=0.46=py_0
opencv=4.0.1=py38h2a7c758_0
openssl=1.1.1i=h8ffe710_0
pandas=1.1.3=py38ha925a31_0
pillow=8.1.0=py38h4fa10fc_0
pip=20.3.3=py38haa95532_0
protobuf=3.14.0=py38h885f38d_1
py-opencv=4.0.1=py38he44ac1e_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycparser=2.20=pyh9f0ad1d_2
pyjwt=2.0.1=pyhd8ed1ab_0
pyopenssl=20.0.1=pyhd8ed1ab_0
pyparsing=2.4.7=pyh9f0ad1d_0
pyqt=5.9.2=py38ha925a31_4
pysocks=1.7.1=py38haa244fe_3
python=3.8.5=h5fd99cc_1
python-dateutil=2.8.1=py_0
python_abi=3.8=1_cp38
pytorch=1.7.1=py3.8_cuda102_cudnn7_0
pytz=2020.1=py_0
qt=5.9.7=vc14h73c81de_0
requests=2.25.1=pyhd3deb0d_0
requests-oauthlib=1.3.0=pyh9f0ad1d_0
rsa=4.7=pyhd3deb0d_0
scikit-learn=0.23.2=py38h47e9c7a_0
scipy=1.5.2=py38h14eb087_0
setuptools=51.3.3=py38haa95532_4
sip=4.19.13=py38ha925a31_0
six=1.15.0=py_0
sqlite=3.33.0=h2a8f88b_0
tensorboard=2.4.1=pyhd8ed1ab_0
tensorboard-plugin-wit=1.8.0=pyh44b312d_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=he774522_0
torchaudio=0.7.2=py38
torchvision=0.8.2=py38_cu102
tornado=6.1=py38h294d835_1
typing-extensions=3.7.4.3=0
typing_extensions=3.7.4.3=py_0
urllib3=1.26.2=pyhd8ed1ab_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
werkzeug=1.0.1=pyh9f0ad1d_0
wheel=0.36.2=pyhd3eb1b0_0
win_inet_pton=1.1.0=py38haa244fe_2
wincertstore=0.2=py38_0
xz=5.2.5=h62dcd97_0
yarl=1.6.3=py38h294d835_1
zipp=3.4.0=py_0
zlib=1.2.11=h62dcd97_4
zstd=1.4.4=ha9fde0e_3
